Move Interaction Result Request to InteractionResultActivity
But because of async is not implemented perfectly, the code is not working as it is made to do
